Makes it easier to combine multiple elements into the scene
Modified my framework to have everything be generated by functions instead of free-form code.  Objects are then pushed to a global array of objects, which are added onto the scene.  This means that when we want to combine our elements, we just need to push what we've made to the global array.